requirements.txt
pyyaml>=6.0
resampy>=0.4.3
numpy>=1.22
paho-mqtt>=2.1.0
PyGObject>=3.42.1<=3.43
torch>=2.4.0
vosk>=0.3.45
